requirements.txt
Django==1.9.5
argparse==1.2.1
dj-static==0.0.6
django-allauth==0.28.0
django-tinymce==2.4.0
gunicorn==19.6.0
httplib2==0.9.2
oauthlib==2.0.0
psycopg2==2.6.2
python-openid==2.2.5
pytz==2016.10
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
static3==0.7.0
twilio==5.6.0
whitenoise==3.2.2
wsgiref==0.1.2
